+++ types/console.d.ts
@@ -2,8 +2,8 @@
  * @Author: BATU1579
  * @CreateDate: 2022-05-25 17:07:59
  * @LastEditor: BATU1579
- * @LastTime: 2022-09-11 10:56:22
- * @FilePath: \\src\\types\\console.d.ts
+ * @LastTime: 2022-10-28 15:51:42
+ * @FilePath: \\types\\console.d.ts
  * @Description: 控制台模块
  */
 
@@ -218,9 +218,9 @@ declare module 'console' {
 
             /**
              * @description: 向控制台输出信息，相当于 `console.log(text)` 。
-             * @param {Object} text 要打印到控制台的信息。
+             * @param {unknown} text 要打印到控制台的信息。
              */
-            print(text: Object): void;
+            print(text: unknown): void;
         }
 
         // 声明全局函数
@@ -246,9 +246,9 @@ declare module 'console' {
 
         /**
          * @description: 向控制台输出信息，相当于 `log(text)` 。
-         * @param {string} text 要打印到控制台的信息。
+         * @param {unknown} text 要打印到控制台的信息。
          */
-        function print(text: Object): void;
+        function print(text: unknown): void;
     }
 
     interface LogConfig {